+++ include/util/mmap_tar.hpp
@@ -15,7 +15,7 @@ namespace osrm::util
 using DataRange = std::pair<const char *, const char *>;
 using DataMap = std::unordered_map<std::string, DataRange>;
 
-inline DataMap mmapTarFile(const boost::filesystem::path &path,
+inline DataMap mmapTarFile(const std::filesystem::path &path,
                            boost::iostreams::mapped_file_source &region)
 {
     DataMap map;