+++ unit_tests/extractor/name_table.cpp
@@ -3,11 +3,11 @@
 
 #include "../common/temporary_file.hpp"
 
-#include <boost/filesystem.hpp>
 #include <boost/test/unit_test.hpp>
 
 #include <iomanip>
 #include <iostream>
+#include <filesystem>
 #include <sstream>
 #include <typeinfo>
 #include <vector>